% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kosmic_math.R
\name{dboxcox}
\alias{dboxcox}
\alias{pboxcox}
\alias{qboxcox}
\alias{rboxcox}
\title{Box-Cox Transformed Normal Distribution}
\usage{
dboxcox(x, mean, sd, lambda)

pboxcox(q, mean, sd, lambda, lower.tail = TRUE)

qboxcox(p, mean, sd, lambda, lower.tail = TRUE)

rboxcox(n, mean, sd, lambda)
}
\description{
Box-Cox Transformed Normal Distribution
}
