% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{call_kosmic_alg}
\alias{call_kosmic_alg}
\title{Call Kosmic algorithm}
\usage{
call_kosmic_alg(
  inputVec,
  n,
  decimals,
  bootstrap,
  bootstrap_seed,
  threads,
  t1min,
  t1max,
  t2min,
  t2max,
  sd,
  tol
)
}
\description{
Call Kosmic algorithm
}
