% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hemoglobin}
\alias{hemoglobin}
\title{Simulated Haemoglobin Results}
\format{
A data frame with columns \code{result}, \code{age} and \code{sex}.
  The result column is numeric and the units are g/dL. The age column is
  numeric and the units are years. The sex column is a character vector,
  containing "m" and "w" for men and women.
}
\source{
\url{https://gitlab.miracum.org/kosmic/kosmic}
}
\usage{
hemoglobin
}
\description{
A simulated hemoglobin result dataset with a true reference
  interval of 12.0-16.0 g/dL.
}
\keyword{datasets}
