% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kosmic-package.R
\docType{package}
\name{kosmic-package}
\alias{kosmic}
\alias{kosmic-package}
\title{kosmic: What the Package Does in One 'Title Case' Line}
\description{
One paragraph description of what the package does as one
        or more full sentences.
}
\details{
A package to estimate the distribution of reference values from a mix of
physiological and abnormal results.
}
