% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kosmic.R
\name{kosmic}
\alias{kosmic}
\alias{kosmic.default}
\alias{kosmic.numeric}
\title{Estimate a Distribution of Physiological Results Using Kosmic}
\usage{
kosmic(data, ...)

\method{kosmic}{default}(data, ...)

\method{kosmic}{numeric}(
  data,
  decimals,
  t1min = 0.05,
  t1max = 0.3,
  t2min = 0.7,
  t2max = 0.95,
  sd_guess = 0.8,
  abstol = 1e-07,
  na.rm = FALSE,
  ...
)
}
\arguments{
\item{data}{A numeric vector. A mixed distribution of physiological and
abnormal results.}

\item{decimals}{A positive integer. The number of digits of precision to
calculate the results to. Increasing this makes the algorithm take longer.}

\item{t1min}{A quantile. Start of the search range for T1.}

\item{t1max}{A quantile. End of the search range for T1.}

\item{t2min}{A quantile. Start of the search range for T2.}

\item{t2max}{A quantile. End of the search range for T1.}

\item{sd_guess}{A quantile. The quantile used for the initial guess of the
standard deviation.}

\item{na.rm}{Logical. If true, any NA and NaNs are removed from \code{data} before
calling kosmic.}

\item{tol}{The absolute convergence tolerance for the optimizer. The
algorithm stops if it is unable to reduce the cost by more than this
amount.}
}
\value{
\code{kosmic} returns an object of class "kosmic". This contains the
following components:

\code{n} The number of data points used to estimate the distribution.

\code{data} A frequency table of the original data, with columns
\code{result} and \code{n}.

\code{estimates} A named vector of estimates for the ditribution:
\code{lambda}, \code{mean}, \code{sd}, \code{t1} and \code{t2}

\code{settings} A named vector of the settings passed to \code{kosmic}.
}
\description{
Estimates the distribution of physiological results from a mixed
distribution of physiological and abnormal results, such as those found in
laboratory databases.
}
\examples{
set.seed(1)
k <- kosmic(haemoglobin$result, 1)
quantile(k)

}
