% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kosmic_bootstrap.R
\name{kosmic_bootstrap}
\alias{kosmic_bootstrap}
\alias{kosmic_bootstrap.default}
\alias{kosmic_bootstrap.numeric}
\title{Bootstrap Estimates of a Distribution of Physiological Results Using Kosmic}
\usage{
kosmic_bootstrap(data, ...)

\method{kosmic_bootstrap}{default}(data, ...)

\method{kosmic_bootstrap}{numeric}(
  data,
  decimals,
  replicates,
  t1min = 0.05,
  t1max = 0.3,
  t2min = 0.7,
  t2max = 0.95,
  sd_guess = 0.8,
  abstol = 1e-07,
  na.rm = FALSE,
  ...
)
}
\description{
Bootstrap Estimates of a Distribution of Physiological Results Using Kosmic
}
