% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kosmic.R
\name{kosmic_bridge}
\alias{kosmic_bridge}
\title{Run Kosmic and Create an Object to Hold the Results}
\usage{
kosmic_bridge(
  data,
  decimals,
  t1min = 0.05,
  t1max = 0.3,
  t2min = 0.7,
  t2max = 0.95,
  sd_guess = 0.8,
  abstol = 1e-07,
  threads = 1L
)
}
\arguments{
\item{data}{A numeric vector.}

\item{decimals}{A positive integer. The number of digits of precision to
calculate the results to. Increasing this makes the algorithm take longer.}

\item{t1min}{A quantile. Start of the search range for T1.}

\item{t1max}{A quantile. End of the search range for T1.}

\item{t2min}{A quantile. Start of the search range for T2.}

\item{t2max}{A quantile. End of the search range for T1.}

\item{sd_guess}{A quantile. The quantile used for the initial guess of the
standard deviation.}

\item{threads}{A positive integer >= 1. The number of threads to use. Default
value is 1. Setting this to a higher number may speed up computation, but
not always. If a value higher than 1 is used the results may not be
reproduceable from run-to-run.}

\item{tol}{The absolute convergence tolerance for the optimizer. The
algorithm stops if it is unable to reduce the cost by more than this
amount.}
}
\value{
A `kosmic` object.
}
\description{
Run Kosmic and Create an Object to Hold the Results
}
