% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{kosmic_impl}
\alias{kosmic_impl}
\title{Call kosmic algorithm}
\usage{
kosmic_impl(
  input_vector,
  decimals,
  bootstrap,
  t1min,
  t1max,
  t2min,
  t2max,
  sd,
  tol
)
}
\description{
Call kosmic algorithm
}
\keyword{internal}
