% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kosmic_graph.R
\name{kosmic_plot_data}
\alias{kosmic_plot_data}
\title{Data to Easily Plot a Distributed Estimated by Kosmic}
\usage{
kosmic_plot_data(k)
}
\arguments{
\item{k}{}
}
\value{

}
\description{
Data to Easily Plot a Distributed Estimated by Kosmic
}
