% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{kosmic_resample}
\alias{kosmic_resample}
\title{Resample data using the same code as the Kosmic agorithm}
\usage{
kosmic_resample(results, counts, decimals)
}
\description{
Resample data using the same code as the Kosmic agorithm
}
