% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kosmic_bootstrap.R
\name{kosmic_resamples}
\alias{kosmic_resamples}
\title{Resamples of a Kosmic Bootstrap}
\usage{
kosmic_resamples(object)
}
\arguments{
\item{object}{}
}
\value{

}
\description{
This function takes an object created by \code{kosmic_bootstrap}
  and returns the resampled data used to generate its results.
}
