% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{kosmic_resamples_impl}
\alias{kosmic_resamples_impl}
\title{Resample data using the same code as the Kosmic agorithm}
\usage{
kosmic_resamples_impl(results, counts, replicates, settings)
}
\description{
Resample data using the same code as the Kosmic agorithm
}
