% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kosmic_graph.R
\name{plot.kosmic}
\alias{plot.kosmic}
\alias{autoplot.kosmic}
\title{Plot a Distribution Estimated by Kosmic}
\usage{
\method{plot}{kosmic}(x, type = c("frequency", "cumulative"), ...)

\method{autoplot}{kosmic}(object, type = c("frequency", "cumulative"), ...)
}
\arguments{
\item{x}{}

\item{type}{}
}
\description{
Plot a Distribution Estimated by Kosmic
}
