% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kosmic.R
\name{quantile.kosmic}
\alias{quantile.kosmic}
\title{Calculate Quantile for an Estimated Distribution of Physiological Results}
\usage{
\method{quantile}{kosmic}(x, probs = c(0.025, 0.5, 0.975), names = TRUE, ...)
}
\arguments{
\item{x}{A kosmic result.}

\item{probs}{Numeric vector. Probabilities with a value between 0 and 1
(exclusive).}

\item{names}{Logical. If true, the result has a names attribute. Set to false
to speed up the calculation of many probabilities.}

\item{...}{}
}
\value{
A numeric vector of the quantiles corresponding to the 2.5th, 50th and 97.5th percentile, or the
probabilities given.
}
\description{
Calculate Quantile for an Estimated Distribution of Physiological Results
}
