% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_seed.R
\name{set_kosmic_seed}
\alias{set_kosmic_seed}
\title{Random Number Generation for Kosmic}
\usage{
set_kosmic_seed(seed)
}
\arguments{
\item{seed}{An single positive integer.}
}
\description{
To make your kosmic results reproduceable from run-to-run, set
the random number generator seed.
}
\examples{
set_kosmic_seed(42)
}
