% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kosmic.R
\name{summary.kosmic}
\alias{summary.kosmic}
\alias{print.summary.kosmic}
\title{Summarising Kosmic Objects}
\usage{
\method{summary}{kosmic}(object, probs = c(0.025, 0.5, 0.975), ...)

\method{print}{summary.kosmic}(x, ...)
}
\arguments{
\item{object}{an object of class "kosmic", usually a result of a call to \link[=kosmic]{kosmic}.}

\item{probs}{a numeric vector, the quantiles of the estimated distribution to be reported
in the summary. Probabilities with a value between 0 and 1 (exclusive).}

\item{...}{further arguments passed to or from other methods.}
}
\value{
The function \code{summary.kosmic} computes and returns
}
\description{
Summarising Kosmic Objects
}
