% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidykosmic-package.R
\docType{package}
\name{tidykosmic-package}
\alias{tidykosmic-package}
\alias{_PACKAGE}
\title{Estimate Reference Intervals from Routinely Collected Laboratory Data}
\description{
To learn more about Kosmic, start with \code{\link{vignette("tidykosmic")}}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://www.divinenephron.co.uk/tidykosmic/}
  \item \url{https://github.com/divinenephron/tidykosmic}
  \item Report bugs at \url{https://github.com/divinenephron/tidykosmic/issues}
}

}
\author{
\strong{Maintainer}: Devon Buchanan \email{devon.buchanan@nhs.net} (Author of the R bindings)

Authors:
\itemize{
  \item Jakob Zierck \email{jakob.zierk@uk-erlangen.de} (Author of the Kosmic C++ library)
  \item Farhad Arzideh \email{farhad.arzideh@uk-koeln.de} (Creator of original concept and prior implementations)
}

}
\keyword{internal}
