% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kosmic_math.R
\name{pboxcox}
\alias{pboxcox}
\alias{qboxcox}
\alias{rboxcox}
\title{Box-Cox Transformed Normal Distribution}
\usage{
pboxcox(q, mean = 0, sd = 1, lambda = 1, lower.tail = TRUE)

qboxcox(p, mean = 0, sd = 1, lambda = 1, lower.tail = TRUE)

rboxcox(n, mean = 0, sd = 1, lambda = 1)
}
\description{
Box-Cox Transformed Normal Distribution
}
