% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kosmic-package.R
\docType{package}
\name{kosmic-package}
\alias{kosmic-package}
\alias{_PACKAGE}
\title{kosmic: Estimate Reference Intervals from Routinely Collected Laboratory Data}
\description{
Kosmic estimates the distribution of physiological results from a mixture of physiological
             and pathological results, like those that are found in laboratory databases. It does this
             by appying a Box-Cox transformation, then fitting a gaussian distribution to a truncated
             part of the data. The estimated distribution can be used to calculated a reference interval.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://divinenephron.github.io/kosmic-r}
  \item \url{https://github.com/divinenephron/kosmic-r}
  \item Report bugs at \url{https://github.com/divinenephron/kosmic-r/issues}
}

}
\author{
\strong{Maintainer}: Devon Buchanan \email{devon.buchanan@nhs.net}

Authors:
\itemize{
  \item Jakob Zierck \email{jakob.zierk@uk-erlangen.de} (Author of the Kosmic C++ library)
  \item Farhad Arzideh \email{farhad.arzideh@uk-koeln.de} (Creator of original concept and prior implementations)
}

}
\keyword{internal}
