% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kosmic.R
\name{kosmic}
\alias{kosmic}
\alias{kosmic.numeric}
\title{Estimate a Distribution of Physiological Results Using Kosmic}
\usage{
kosmic(data, ...)

\method{kosmic}{numeric}(
  data,
  decimals,
  t1min = 0.05,
  t1max = 0.3,
  t2min = 0.7,
  t2max = 0.95,
  sd_guess = 0.8,
  abstol = 1e-07,
  threads = 1L,
  na.rm = FALSE,
  ...
)
}
\arguments{
\item{data}{A numeric vector. A mixed distribution of physiological and
abnormal results.}

\item{decimals}{A positive integer. The number of digits of precision to
calculate the results to. Increasing this makes the algorithm take longer.}

\item{t1min}{A quantile. Start of the search range for T1.}

\item{t1max}{A quantile. End of the search range for T1.}

\item{t2min}{A quantile. Start of the search range for T2.}

\item{t2max}{A quantile. End of the search range for T1.}

\item{sd_guess}{A quantile. The quantile used for the initial guess of the
standard deviation.}

\item{threads}{A positive integer >= 1. The number of threads to use. Default
value is 1. Setting this to a higher number may speed up computation, but
not always. If a value higher than 1 is used the results may not be
reproduceable from run-to-run.}

\item{na.rm}{Logical. If true, any NA and NaNs are removed from `data` before
calling kosmic.}

\item{tol}{The absolute convergence tolerance for the optimizer. The
algorithm stops if it is unable to reduce the cost by more than this
amount.}
}
\value{
`kosmic` returns an object of class "kosmic". This contains the
  following components:

  \code{n} The number of data points used to estimate the distribution.

  \code{estimates} A named vector of estimates for the ditribution:
  \code{lambda}, \code{mean}, \code{sd}.

  \code{truncation} A named vector of estimate for truncating the original
  data: \code{t1} and \code{t2}.

  \code{settings} A named vector of the settings passed to \code{kosmic}.
}
\description{
Estimates the distribution of physiological results from a mixed
  distribution of physiological and abnormal results, such as those found in
  laboratory databases.
}
\details{
The kosmic algorithm works by minimizes the difference between an
  estimated parametric distribution and a truncated part of the observed
  distribution, specifically, the Kolmogorov-Smirnov-distance between a
  hypothetical Gaussian distribution and the observed distribution of test
  results after Box-Cox-transformation.
}
\examples{
set.seed(1)
k <- kosmic(rnorm(10000, 16, 1), 1)
quantile(k)

}
