% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kosmic.R
\name{kosmic_bridge}
\alias{kosmic_bridge}
\title{Run kosmic and Create an Object to Hold the Results}
\usage{
kosmic_bridge(data, decimals, t1min, t1max, t2min, t2max, sd_guess, abstol)
}
\value{
A \code{kosmic} object.
}
\description{
Run kosmic and Create an Object to Hold the Results
}
\keyword{internal}
