% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kosmic.R
\name{new_kosmic}
\alias{new_kosmic}
\title{Create a new kosmic result}
\usage{
new_kosmic(
  n,
  lambda,
  mean,
  sd,
  t1,
  t2,
  decimals,
  t1min,
  t1max,
  t2min,
  t2max,
  sd_guess,
  abstol
)
}
\arguments{
\item{n}{A positive number. The number of results used to estimate the
distribution.}

\item{lambda}{A number. The lambda parameter of the Box-Cox transformation for the
estimated distribution. It which describes the skewness.}

\item{mean}{A number. The mean parameter shows the central point of the
estimated distribution before it undergoes Box-Cox transformation.}

\item{sd}{A number. The spread of the estimated distribution before it
undergoes Box-Cox transformation.}

\item{t1}{A number.}

\item{t2}{A number.}

\item{decimals}{A number. The number of digits of precision the kosmic
algorithm was configured for.}

\item{t1min}{A number.}

\item{t1max}{A number.}

\item{t2min}{A number.}

\item{t2max}{A number.}

\item{sd_guess}{A number.}

\item{abstol}{A number.}
}
\value{
A list with the class `kosmic`, containing the runtime parameters and results
of the kosmic algorithm.
}
\description{
This creates an S3 object to hold the results of the kosmic library being run
on a set of data. It should be called by user-facing methods such as
`kosmic()`.
}
