% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kosmic_graph.R
\name{autoplot.kosmic}
\alias{autoplot.kosmic}
\title{Plot a Distribution Estimated by Kosmic}
\usage{
\method{autoplot}{kosmic}(object, type = c("frequency", "cumulative"))
}
\arguments{
\item{object}{}

\item{type}{}
}
\description{
Plot a Distribution Estimated by Kosmic
}
